Minimal port mapping
Using --link gives a container access to another containers exposed ports, so a port mapping is not necessary. Since most people are not interested in the data inside the database itself, we can remove the port mapping and avoid problems if the port is already used by a different mongo instance. (https://docs.docker.com/engine/reference/run/#expose-incoming-ports)

I also updated the mongo-container name and link to reduce the risk of name clashes on another devs PC.