Fix improper shutdown
Run strapi in the background so we can catch signals and properly shutdown strapi.
Docker stop sends a SIGTERM which is ignored by strapi and causes Docker to kill the container (exit 137). This catches the SIGTERM and sends a SIGINT (CTRL-C) to strapi.
We cannot use STOPSIGNAL in the Dockerfile as SIGINT is a signal that will not be passed to PID 1 (which is the wrapper-script).